+++ CHANGELOG.md
@@ -7,5 +7,7 @@
 - Added (ᵔᴥᵔ) as a trigger for the :dog: emoji
 
 ### Fixes
+- Fixed sea creature glow only displaying one entity and invisible entities sometimes
 
 ### Back-end
+- Reworked entity system to accept more model entities